move encoding layers to Layers.encodings
